refactor: WinningResult 패키지 위치 수정
